Subject: EditSymptomActivity
What Happened: Implemented EditSymptomActivity to do an update of Symptons in the Symptoms table and a cascade update of Symptoms in the SymTrax Table. Implemented EditSymptomActivity to do a delete of Symptoms in the Symptoms table and a Dialog trying to explain why a Symptom can not be deleted in the SymTrax table (NO cascade deletes allowed).